mongo_to_bigquery/requirements.txt
pandas
google-cloud-bigquery
python-dotenv


